Updated the InventoryReportController to show inventory of stocks from a given date (with pagination).
Updated the StockController to show count of stocks, and fixed exceptions in handling dynamic pages in the combobox.

Updated the StockEntryController to handle critical quantity, disable other fields when adding existing stock to entry logs, and added reset fields function.

Updated the TrashStockController to include count of trashed stocks and fixed minor bug in handling pagination.